+++ src/components/lineworks/index.ts
@@ -4,6 +4,5 @@
  * @module components/lineworks
  */
 
-export { default as LineWorksSettingsModal } from './LineWorksSettingsModal';
 export { default as LineWorksSendButton } from './LineWorksSendButton';
 export { default as LineWorksPreviewModal } from './LineWorksPreviewModal';